dom/base/test/test_bug1648887.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1648887
-->
<head>
<title>Test for Bug 1648887</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1648887">Mozilla Bug 1648887</a>
<p id="display"></p>
<iframe srcdoc="<a id='a' href='http://w'>"></iframe>
<script class="testbody" type="text/javascript">

/** Test for Bug 1648887 **/
add_task(async function test() {
  await SpecialPowers.pushPrefEnv({
    set: [
      ["dom.disable_open_during_load", false],
    ],
  });

  let iframe = document.querySelector("iframe");
  iframe.contentDocument.getElementById('a').click();
  iframe.contentWindow.open(document.createElement('r').outerHTML,'','h').close();
  ok(true, "Should not crash");
});

</script>
</pre>
</body>
</html>
